requirements.txt
langchain
langchain_openai
langgraph
pillow
openai
python-dotenv
torch